Rename accessible module to object_reference
