+++ atspi-common/src/lib.rs
@@ -13,8 +13,8 @@ extern crate static_assertions;
 #[macro_use]
 pub(crate) mod macros;
 
-pub mod accessible;
-pub use accessible::Accessible;
+pub mod object_reference;
+pub use object_reference::ObjectReference;
 pub mod interface;
 pub use interface::{Interface, InterfaceSet};
 pub mod state;
